tiled-ai-tagger/server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
pydantic==2.6.1
pydantic-settings==2.1.0 